test(storage): refactor unit tests for public & internal facade
