+++ packages/storage/__tests__/providers/s3/apis/internal/uploadData/index.test.ts
@@ -1,24 +1,27 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { defaultStorage } from '@aws-amplify/core';
-
-import { uploadData } from '../../../../../src/providers/s3/apis';
-import { MAX_OBJECT_SIZE } from '../../../../../src/providers/s3/utils/constants';
-import { createUploadTask } from '../../../../../src/providers/s3/utils';
+import { uploadData } from '../../../../../../src/providers/s3/apis/internal/uploadData';
+import { MAX_OBJECT_SIZE } from '../../../../../../src/providers/s3/utils/constants';
+import { createUploadTask } from '../../../../../../src/providers/s3/utils';
 import {
 	StorageValidationErrorCode,
 	validationErrorMap,
-} from '../../../../../src/errors/types/validation';
-import { putObjectJob } from '../../../../../src/providers/s3/apis/internal/uploadData/putObjectJob';
-import { getMultipartUploadHandlers } from '../../../../../src/providers/s3/apis/internal/uploadData/multipart';
-import { UploadDataInput, UploadDataWithPathInput } from '../../../../../src';
+} from '../../../../../../src/errors/types/validation';
+import { putObjectJob } from '../../../../../../src/providers/s3/apis/internal/uploadData/putObjectJob';
+import { getMultipartUploadHandlers } from '../../../../../../src/providers/s3/apis/internal/uploadData/multipart';
+import {
+	UploadDataInput,
+	UploadDataWithPathInput,
+} from '../../../../../../src';
 
-jest.mock('../../../../../src/providers/s3/utils/');
+jest.mock('../../../../../../src/providers/s3/utils/');
 jest.mock(
-	'../../../../../src/providers/s3/apis/internal/uploadData/putObjectJob',
+	'../../../../../../src/providers/s3/apis/internal/uploadData/putObjectJob',
 );
-jest.mock('../../../../../src/providers/s3/apis/internal/uploadData/multipart');
+jest.mock(
+	'../../../../../../src/providers/s3/apis/internal/uploadData/multipart',
+);
 
 const testPath = 'testPath/object';
 const validBucketOwner = '111122223333';
@@ -249,12 +252,7 @@ describe('uploadData with path', () => {
 
 			expect(mockPutObjectJob).not.toHaveBeenCalled();
 			expect(mockGetMultipartUploadHandlers).toHaveBeenCalledWith(
-				expect.objectContaining({
-					...testInput,
-					options: {
-						resumableUploadsCache: defaultStorage,
-					},
-				}),
+				expect.objectContaining(testInput),
 				expect.any(Number),
 			);
 		});
